requirements.txt
black==23.3.0
click==8.1.3
et-xmlfile==1.1.0
mypy-extensions==1.0.0
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
